test_viewsをCRUDとバリデーションに分割
